# docs/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Bardic server API
  version: 0.2.6
  summary: Contract between the Bardic server and its clients (draft).
  description: |
    Normative, language-neutral contract for Bardic v2. It is derived from
    `docs/PRODUCT-SPEC.md`, not from the prototype. Clients are generated
    from this file; the server must satisfy it. Read `README.md` in this folder
    first: it explains identity headers, errors, money, events, long-running
    work and the place-sync protocol once, so each operation does not repeat them.

    Status: pre-release (0.x). Until 1.0.0, additive changes bump the patch
    version and breaking changes bump the minor version.
  license:
    name: Personal use
    identifier: LicenseRef-Personal
servers:
  - url: http://localhost:8765
    description: Default local address. The server name and address are configurable.
...
paths:
...
  /api/audiobooks/{audiobook_id}/exports:
    parameters: [{$ref: '#/components/parameters/AudiobookId'}]
    post:
      operationId: createExport
      tags: [Admin]
      summary: Export an audiobook as one audio file with chapters
      description: |
        Starts a job that encodes the ready chapters into one M4B with chapter
        markers, using `ffmpeg` on the Bardic computer. Chapters not made yet are
        left out. Errors (409): `nothing_ready`, `encoder_missing` (no ffmpeg).
        A failed export has job state `failed` with `needs_you.code`
        `export_failed`. `downloadExport` is 409 `export_not_ready` until the
        export is `ready`, and supports `Range`.
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                format: {type: string, enum: [m4b], default: m4b}
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '202':
          description: The export.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Export'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/exports/{export_id}:
    parameters:
      - {name: export_id, in: path, required: true, schema: {type: string}}
    get:
      operationId: getExport
      tags: [Admin]
      summary: Read an export
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: The export.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Export'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/exports/{export_id}/file:
    parameters:
      - {name: export_id, in: path, required: true, schema: {type: string}}
    get:
      operationId: downloadExport
      tags: [Admin]
      summary: Download the exported file
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: The file.
          content:
            audio/mp4:
              schema: {type: string, format: binary}
        '206':
          description: Partial file (the request had a `Range` header).
          content:
            audio/mp4:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  parameters:
...
    Device:
      name: X-Bardic-Device
      in: header
      required: true
      description: Client-generated device id (UUID). Registers the device on first use and is recorded in the audit log.
      schema: {type: string}
...
    AudiobookId: {name: audiobook_id, in: path, required: true, schema: {type: string}}
...
  responses:
    InvalidRequest:
      description: The request is malformed or breaks a rule; `code` says which.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: The resource does not exist (or is not visible to this listener).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: The request conflicts with current state; `code` says how.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    ServerError:
      description: An unexpected failure on the server. Nothing you need to undo was changed; `retryable` is true.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
...
    Error:
      type: object
      required: [code, detail]
      properties:
        code: {type: string, description: "Stable machine-readable code (open enumeration); see README."}
        detail: {type: string, description: "Human-readable, safe to show."}
        retryable: {type: boolean}
        retry_after_seconds: {type: [integer, 'null']}
        context:
          type: object
          additionalProperties: true
          description: Code-specific fields (for example `plan_id`, `existing_book_id`).
...
    Export:
      type: object
      required: [id, audiobook_id, state, format, bytes, job_id, created_at]
      properties:
        id: {type: string}
        audiobook_id: {type: string}
        state: {type: string, enum: [queued, running, ready, failed]}
        format: {type: string, enum: [m4b]}
        bytes: {type: [integer, 'null']}
        job_id: {type: string}
        created_at: {type: string, format: date-time}
...